added routes diagram tp readme
