+++ src/controllers/views.js
@@ -1,4 +1,6 @@
-const renderHtml = (req, res) => {};
+//functions to add a front end part - to be developped in the next update
+const renderStartPage = (req, res) => {};
+const renderReportPage = (req, res) => {};
 const renderErrorPage = (req, res) => {};
 
-module.exports = { renderHtml };
+module.exports = { renderStartPage, renderReportPage, renderErrorPage };
