+++ src/routes/api/departments.js
@@ -4,11 +4,13 @@ const {
   addDepartment,
   removeDepartment,
   viewDepartmentSpend,
+  viewEmployeesByDepartment,
 } = require("../../controllers/data/department");
 
 const router = Router();
 
 router.get("/", viewAllDepartments);
+router.get("/employees/:id", viewEmployeesByDepartment);
 router.post("/", addDepartment);
 router.delete("/:id", removeDepartment);
 router.get("/spend/:id", viewDepartmentSpend);
