+++ src/routes/api/employees.js
@@ -3,7 +3,6 @@ const {
   viewAllEmployees,
   addEmployee,
   removeEmployee,
-  viewEmployeesByDepartment,
   updateEmployee,
 } = require("../../controllers/data/employees");
 
@@ -13,6 +12,5 @@ router.get("/", viewAllEmployees);
 router.post("/", addEmployee);
 router.delete("/:id", removeEmployee);
 router.put("/:id", updateEmployee);
-router.get("/department/:id", viewEmployeesByDepartment);
 
 module.exports = router;
